% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_FOUNDER_COLORS}
\alias{DEFAULT_FOUNDER_COLORS}
\title{Default founder colors for DO/CC mice visualization}
\format{
A named character vector with hex color codes:
\describe{
  \item{A}{#F0F000 - Yellow (A/J)}
  \item{B}{#808080 - Gray (C57BL/6J)}
  \item{C}{#F08080 - Light Red (129S1/SvImJ)}
  \item{D}{#1010F0 - Blue (NOD/ShiLtJ)}
  \item{E}{#00A0F0 - Light Blue (NZO/H1LtJ)}
  \item{F}{#00A000 - Green (CAST/EiJ)}
  \item{G}{#F00000 - Red (PWK/PhJ)}
  \item{H}{#9000E0 - Purple (WSB/EiJ)}
}
}
\usage{
DEFAULT_FOUNDER_COLORS
}
\description{
Standard color scheme for founder strains in genome plots.
Colors are optimized for publication and accessibility.
}
\keyword{datasets}
