% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbrsR.R
\docType{data}
\name{DEFAULT_FOUNDERS}
\alias{DEFAULT_FOUNDERS}
\title{Default founder strains for DO/CC mice}
\format{
A character vector with 8 founder strains:
\describe{
  \item{A}{A/J}
  \item{B}{C57BL/6J}
  \item{C}{129S1/SvImJ}
  \item{D}{NOD/ShiLtJ}
  \item{E}{NZO/H1LtJ}
  \item{F}{CAST/EiJ}
  \item{G}{PWK/PhJ}
  \item{H}{WSB/EiJ}
}
}
\usage{
DEFAULT_FOUNDERS
}
\description{
Standard founder strain names used in Diversity Outbred (DO) and 
Collaborative Cross (CC) mouse studies.
}
\keyword{datasets}
