% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{generate_diplotypes}
\alias{generate_diplotypes}
\title{Generate all possible founder strain combinations}
\usage{
generate_diplotypes(founders = DEFAULT_FOUNDERS)
}
\arguments{
\item{founders}{Character vector of founder strain names}
}
\value{
Character vector of all possible diplotype combinations
}
\description{
Creates all possible diploid combinations from founder strain names.
This is used internally by load_npz_file to generate column names.
}
\examples{
generate_diplotypes(c('A', 'B', 'C'))
# Returns: c('AA', 'AB', 'AC', 'BA', 'BB', 'BC', 'CA', 'CB', 'CC')

}
